pyproject.toml
[project]
name = "banner-pipeline"
version = "0.1.0"
description = "Video banner/logo replacement pipeline using SAM2 segmentation"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.8",
    "numpy>=1.24",
    "matplotlib>=3.7",
    "Pillow>=10.0",
    "scipy>=1.11",
    "pyyaml>=6.0",
    "torch>=2.0",
    "torchvision>=0.15",
    "pytest>=7.0",
    "ruff>=0.11",
    "mypy>=1.15",
    "pre-commit>=4.0",
    "types-PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/banner_pipeline"]

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
